+++ src/iife.ts
@@ -10,7 +10,8 @@
 
 import { connect } from "./connect.js";
 import { downloadFile } from "./download-file.js";
-import { action, pickFile, pickFiles } from "./extensions.js";
+import { HostCapabilityError } from "./errors.js";
+import { action, hostSupports, pickFile, pickFiles } from "./extensions.js";
 import { connectUI } from "./host/connect.js";
 import { callToolAsTask } from "./task-handle.js";
 
@@ -20,6 +21,8 @@ import { callToolAsTask } from "./task-handle.js";
   callToolAsTask,
   action,
   downloadFile,
+  hostSupports,
   pickFile,
   pickFiles,
+  HostCapabilityError,
 };